requirements.txt
beautifulsoup4==4.9.3
fastapi==0.79.0
Jinja2==3.1.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
pydantic==1.9.2
python-dotenv==0.20.0
requests==2.28.1
uvicorn==0.18.2
httpx==0.24.1
